[ui/calendar] Update: CalendarAdapter -> CalendarLayout dependency
- Update: CalendarAdapter.view checks layout type
- Add: CalendarLayout.DataSet
- Update: CalendarLayout.invalidLayoutIfNeeded(dataSet:)

Although it seems like a better solution to have CalendarLayout observe properties of CalendarAdapter,
there is no good way to observe the properties of CalendarAdapter without using Rx or making CalendarAdapter an NSObject.

When the minimum target is changed so that Combine can be used, CalendarLayout may be refactored,
but since CalendarAdapter has to assert the layout instance type of `view` anyways,
and the data parameters needed for the layout is not likely to change,
we'll go along with having CalendarAdapter explicitly call for invalidation.